# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: AI 智能下单与生产协同平台 API
  version: 1.0.0
  description: |
    牙科定制工厂三端系统（医生端 / 客服端 / 生产端 / 管理端）后端 API 规范。

    ## 认证方式
    所有接口（除登录外）必须携带 Authorization: Bearer {access_token}。
    服务端对每个接口强制校验角色权限，禁止仅靠前端隐藏。

    ## 角色枚举
    - DOCTOR：医生
    - CS：客服
    - WORKER：技工/生产人员
    - ADMIN：超级管理员

    ## 通用响应格式
    所有接口统一返回：
    { "code": 200, "msg": "success", "data": {} }
    错误时：
    { "code": 4xx/5xx, "msg": "错误说明", "data": null }
  license:
    name: Proprietary internal API
    url: https://yourdomain.com/legal/proprietary-api-license
servers:
- url: https://api.yourdomain.com/v1
  description: 正式环境
- url: https://test-api.yourdomain.com/v1
  description: 测试环境
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    CommonResponse:
      type: object
      properties:
        code:
          type: integer
          example: 200
        msg:
          type: string
          example: success
        data: {}
...
    MessageItem:
      type: object
      properties:
        msg_id:
          type: integer
        order_id:
          type: integer
        order_no:
          type: string
          description: 订单号，供客服待审核队列和订单消息上下文识别订单。
        product_type:
          type: string
          description: 产品类型。
        external_status:
          type: string
          description: 医生端外部状态。
        sender_user_id:
          type: integer
        sender_role:
          type: string
          enum:
          - DOCTOR
          - CS
          - WORKER
        sender_name:
          type: string
        content:
          type: string
        edited_content:
          type: string
          description: 客服编辑后内容
        attachment_urls:
          type: array
          items:
            type: string
        visible_to:
          type: string
        review_status:
          "$ref": "#/components/schemas/MessageStatus"
        review_note:
          type: string
        create_time:
          type: string
          format: date-time
...
    MessageRequest:
      type: object
      required:
      - content
      properties:
        content:
          type: string
        attachment_file_ids:
          type: array
          items:
            type: integer
        visible_to:
          type: string
          enum:
          - DOCTOR
          - DOCTOR_CS
          - INTERNAL
          - ALL
    MessageReviewRequest:
      type: object
      required:
      - action
      properties:
        action:
          type: string
          enum:
          - APPROVE
          - REJECT
        edited_content:
          type: string
        review_note:
          type: string
...
  responses:
    BadRequest:
      description: 请求参数错误或校验失败
      content:
        application/json:
          schema:
            "$ref": "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: 未登录或 Token 无效
      content:
        application/json:
          schema:
            "$ref": "#/components/schemas/ErrorResponse"
    Forbidden:
      description: 无权限或数据范围拒绝
      content:
        application/json:
          schema:
            "$ref": "#/components/schemas/ErrorResponse"
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema:
            "$ref": "#/components/schemas/ErrorResponse"
    Conflict:
      description: 业务状态冲突或幂等冲突
      content:
        application/json:
          schema:
            "$ref": "#/components/schemas/ErrorResponse"
    ServiceUnavailable:
      description: 外部依赖或存储服务不可用
      content:
        application/json:
          schema:
            "$ref": "#/components/schemas/ErrorResponse"
    DefaultError:
      description: 未分类错误响应
      content:
        application/json:
          schema:
            "$ref": "#/components/schemas/ErrorResponse"
security:
- BearerAuth: []
paths:
...
  "/orders/{orderId}/messages":
    get:
      tags:
      - Message
      summary: 获取订单消息列表
      description: |
        权限：DOCTOR / CS / WORKER / ADMIN
        数据层按角色过滤：
        - DOCTOR：只返回 review_status=DIRECT 或 APPROVED 的消息，且 visible_to 包含 DOCTOR
        - WORKER：只返回 visible_to 包含 WORKER 的消息
        - CS / ADMIN：返回全部消息（含 PENDING_REVIEW / REJECTED）
      parameters:
      - name: orderId
        in: path
        required: true
        schema:
          type: integer
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                allOf:
                - "$ref": "#/components/schemas/CommonResponse"
                - type: object
                  properties:
                    data:
                      type: array
                      items:
                        "$ref": "#/components/schemas/MessageItem"
        '400':
          "$ref": "#/components/responses/BadRequest"
        '401':
          "$ref": "#/components/responses/Unauthorized"
        '403':
          "$ref": "#/components/responses/Forbidden"
        '404':
          "$ref": "#/components/responses/NotFound"
        '409':
          "$ref": "#/components/responses/Conflict"
        default:
          "$ref": "#/components/responses/DefaultError"
        '503':
          "$ref": "#/components/responses/ServiceUnavailable"
      operationId: getOrdersByOrderIdMessages
    post:
      tags:
      - Message
      summary: 发送消息
      description: |
        权限：DOCTOR / CS / WORKER / ADMIN
        - DOCTOR 发送：review_status=DIRECT，visible_to=CS_ONLY（生产端不可见）
        - CS 发送：review_status=DIRECT，visible_to 由请求参数决定
        - WORKER 发送：review_status=PENDING_REVIEW，visible_to=ALL（审核通过后医生可见）
      parameters:
      - name: orderId
        in: path
        required: true
        schema:
          type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/MessageRequest"
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                allOf:
                - "$ref": "#/components/schemas/CommonResponse"
                - type: object
                  properties:
                    data:
                      "$ref": "#/components/schemas/MessageItem"
        '400':
          "$ref": "#/components/responses/BadRequest"
        '401':
          "$ref": "#/components/responses/Unauthorized"
        '403':
          "$ref": "#/components/responses/Forbidden"
        '404':
          "$ref": "#/components/responses/NotFound"
        '409':
          "$ref": "#/components/responses/Conflict"
        default:
          "$ref": "#/components/responses/DefaultError"
        '503':
          "$ref": "#/components/responses/ServiceUnavailable"
      operationId: postOrdersByOrderIdMessages
  "/messages/{msgId}/review":
    post:
      tags:
      - Message
      summary: 客服审核生产端消息
      description: 权限：CS / ADMIN
      parameters:
      - name: msgId
        in: path
        required: true
        schema:
          type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/MessageReviewRequest"
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                allOf:
                - "$ref": "#/components/schemas/CommonResponse"
                - type: object
                  properties:
                    data:
                      "$ref": "#/components/schemas/MessageItem"
        '400':
          "$ref": "#/components/responses/BadRequest"
        '401':
          "$ref": "#/components/responses/Unauthorized"
        '403':
          "$ref": "#/components/responses/Forbidden"
        '404':
          "$ref": "#/components/responses/NotFound"
        '409':
          "$ref": "#/components/responses/Conflict"
        default:
          "$ref": "#/components/responses/DefaultError"
        '503':
          "$ref": "#/components/responses/ServiceUnavailable"
      operationId: postMessagesByMsgIdReview
  "/messages/pending-review":
    get:
      tags:
      - Message
      summary: 获取待审核消息队列（汇总）
      description: 权限：CS / ADMIN
      parameters:
      - name: page
        in: query
        schema:
          type: integer
          default: 1
      - name: size
        in: query
        schema:
          type: integer
          default: 20
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                allOf:
                - "$ref": "#/components/schemas/CommonResponse"
                - type: object
                  properties:
                    data:
                      type: array
                      items:
                        "$ref": "#/components/schemas/MessageItem"
        '400':
          "$ref": "#/components/responses/BadRequest"
        '401':
          "$ref": "#/components/responses/Unauthorized"
        '403':
          "$ref": "#/components/responses/Forbidden"
        '404':
          "$ref": "#/components/responses/NotFound"
        '409':
          "$ref": "#/components/responses/Conflict"
        default:
          "$ref": "#/components/responses/DefaultError"
        '503':
          "$ref": "#/components/responses/ServiceUnavailable"
      operationId: getMessagesPendingReview
  "/orders/{orderId}/design-drafts":
    get:
      tags:
      - DesignDraft
      summary: 获取订单设计稿列表（所有版本）
      description: |
        权限：DOCTOR / CS / WORKER / ADMIN
        医生端只返回 status 为 PENDING_DOCTOR_CONFIRM / DOCTOR_CONFIRMED / DOCTOR_REJECTED 的版本。
        CS_REJECTED 的版本医生端不可见。
      parameters:
      - name: orderId
        in: path
        required: true
        schema:
          type: integer
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                allOf:
                - "$ref": "#/components/schemas/CommonResponse"
                - type: object
                  properties:
                    data:
                      type: array
                      items:
                        "$ref": "#/components/schemas/DesignDraftResponse"
        '400':
          "$ref": "#/components/responses/BadRequest"
        '401':
          "$ref": "#/components/responses/Unauthorized"
        '403':
          "$ref": "#/components/responses/Forbidden"
        '404':
          "$ref": "#/components/responses/NotFound"
        '409':
          "$ref": "#/components/responses/Conflict"
        default:
          "$ref": "#/components/responses/DefaultError"
        '503':
          "$ref": "#/components/responses/ServiceUnavailable"
      operationId: getOrdersByOrderIdDesignDrafts
    post:
      tags:
      - DesignDraft
      summary: 上传新版设计稿
      description: |
        权限：WORKER / CS / ADMIN（工厂端任意人员）
        任务 9D.13 第一增量支持同一版本绑定多个 file_ids；响应保留 file_id 作为第一个文件的兼容字段。
      parameters:
      - name: orderId
        in: path
        required: true
        schema:
          type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/DesignDraftRequest"
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                allOf:
                - "$ref": "#/components/schemas/CommonResponse"
                - type: object
                  properties:
                    data:
                      "$ref": "#/components/schemas/DesignDraftResponse"
        '400':
          "$ref": "#/components/responses/BadRequest"
        '401':
          "$ref": "#/components/responses/Unauthorized"
        '403':
          "$ref": "#/components/responses/Forbidden"
        '404':
          "$ref": "#/components/responses/NotFound"
        '409':
          "$ref": "#/components/responses/Conflict"
        default:
          "$ref": "#/components/responses/DefaultError"
        '503':
          "$ref": "#/components/responses/ServiceUnavailable"
      operationId: postOrdersByOrderIdDesignDrafts
  "/orders/{orderId}/design-drafts/{draftId}/cs-review":
    post:
      tags:
      - DesignDraft
      summary: 客服审核设计稿
      description: 权限：CS / ADMIN
      parameters:
      - name: orderId
        in: path
        required: true
        schema:
          type: integer
      - name: draftId
        in: path
        required: true
        schema:
          type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/DesignDraftReviewRequest"
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                allOf:
                - "$ref": "#/components/schemas/CommonResponse"
                - type: object
                  properties:
                    data:
                      "$ref": "#/components/schemas/DesignDraftResponse"
        '400':
          "$ref": "#/components/responses/BadRequest"
        '401':
          "$ref": "#/components/responses/Unauthorized"
        '403':
          "$ref": "#/components/responses/Forbidden"
        '404':
          "$ref": "#/components/responses/NotFound"
        '409':
          "$ref": "#/components/responses/Conflict"
        default:
          "$ref": "#/components/responses/DefaultError"
        '503':
          "$ref": "#/components/responses/ServiceUnavailable"
      operationId: postOrdersByOrderIdDesignDraftsByDraftIdCsReview
  "/orders/{orderId}/design-drafts/{draftId}/doctor-confirm":
    post:
      tags:
      - DesignDraft
      summary: 医生确认或驳回设计稿
      description: 权限：DOCTOR（只能操作本人订单的设计稿）
      parameters:
      - name: orderId
        in: path
        required: true
        schema:
          type: integer
      - name: draftId
        in: path
        required: true
        schema:
          type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/DoctorDraftConfirmRequest"
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                allOf:
                - "$ref": "#/components/schemas/CommonResponse"
                - type: object
                  properties:
                    data:
                      "$ref": "#/components/schemas/DesignDraftResponse"
        '400':
          "$ref": "#/components/responses/BadRequest"
        '401':
          "$ref": "#/components/responses/Unauthorized"
        '403':
          "$ref": "#/components/responses/Forbidden"
        '404':
          "$ref": "#/components/responses/NotFound"
        '409':
          "$ref": "#/components/responses/Conflict"
        default:
          "$ref": "#/components/responses/DefaultError"
        '503':
          "$ref": "#/components/responses/ServiceUnavailable"
      operationId: postOrdersByOrderIdDesignDraftsByDraftIdDoctorConfirm
  "/orders/{orderId}/bill":
    get:
      tags:
      - Bill
      summary: 获取账单信息
      description: 权限：DOCTOR（只读）/ CS / ADMIN
      parameters:
      - name: orderId
        in: path
        required: true
        schema:
          type: integer
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                allOf:
                - "$ref": "#/components/schemas/CommonResponse"
                - type: object
                  properties:
                    data:
                      "$ref": "#/components/schemas/BillResponse"
        '400':
          "$ref": "#/components/responses/BadRequest"
        '401':
          "$ref": "#/components/responses/Unauthorized"
        '403':
          "$ref": "#/components/responses/Forbidden"
        '404':
          "$ref": "#/components/responses/NotFound"
        '409':
          "$ref": "#/components/responses/Conflict"
        default:
          "$ref": "#/components/responses/DefaultError"
        '503':
          "$ref": "#/components/responses/ServiceUnavailable"
      operationId: getOrdersByOrderIdBill
    post:
      tags:
      - Bill
      summary: 上传账单文件
      description: 权限：CS / ADMIN
      parameters:
      - name: orderId
        in: path
        required: true
        schema:
          type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/BillRequest"
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                allOf:
                - "$ref": "#/components/schemas/CommonResponse"
                - type: object
                  properties:
                    data:
                      "$ref": "#/components/schemas/BillResponse"
        '400':
          "$ref": "#/components/responses/BadRequest"
        '401':
          "$ref": "#/components/responses/Unauthorized"
        '403':
          "$ref": "#/components/responses/Forbidden"
        '404':
          "$ref": "#/components/responses/NotFound"
        '409':
          "$ref": "#/components/responses/Conflict"
        default:
          "$ref": "#/components/responses/DefaultError"
        '503':
          "$ref": "#/components/responses/ServiceUnavailable"
      operationId: postOrdersByOrderIdBill
  "/orders/{orderId}/bill/payment-status":
    post:
      tags:
      - Bill
      summary: 人工维护付款状态
      description: 权限：CS / ADMIN；一期只维护对外付款状态，不接真实支付系统。
      parameters:
      - name: orderId
        in: path
        required: true
        schema:
          type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/PaymentStatusRequest"
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                allOf:
                - "$ref": "#/components/schemas/CommonResponse"
                - type: object
                  properties:
                    data:
                      "$ref": "#/components/schemas/BillResponse"
        '400':
          "$ref": "#/components/responses/BadRequest"
        '401':
          "$ref": "#/components/responses/Unauthorized"
        '403':
          "$ref": "#/components/responses/Forbidden"
        '404':
          "$ref": "#/components/responses/NotFound"
        '409':
          "$ref": "#/components/responses/Conflict"
        default:
          "$ref": "#/components/responses/DefaultError"
        '503':
          "$ref": "#/components/responses/ServiceUnavailable"
      operationId: postOrdersByOrderIdBillPaymentStatus
  "/orders/{orderId}/payments":
    get:
      tags:
      - Bill
      summary: 获取人工收款流水
      description: 权限：DOCTOR（只读本人订单）/ CS / ADMIN；一期只返回人工维护流水，不接真实支付系统。
      parameters:
      - name: orderId
        in: path
        required: true
        schema:
          type: integer
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                allOf:
                - "$ref": "#/components/schemas/CommonResponse"
                - type: object
                  properties:
                    data:
                      type: array
                      items:
                        "$ref": "#/components/schemas/PaymentRecordResponse"
        '400':
          "$ref": "#/components/responses/BadRequest"
        '401':
          "$ref": "#/components/responses/Unauthorized"
        '403':
          "$ref": "#/components/responses/Forbidden"
        '404':
          "$ref": "#/components/responses/NotFound"
        '409':
          "$ref": "#/components/responses/Conflict"
        default:
          "$ref": "#/components/responses/DefaultError"
        '503':
          "$ref": "#/components/responses/ServiceUnavailable"
      operationId: getOrdersByOrderIdPayments
    post:
      tags:
      - Bill
      summary: 录入人工收款流水
      description: 权限：CS / ADMIN；一期人工录入收款事实，不接真实支付网关、退款、对账或发票。
      parameters:
      - name: orderId
        in: path
        required: true
        schema:
          type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/PaymentRecordRequest"
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                allOf:
                - "$ref": "#/components/schemas/CommonResponse"
                - type: object
                  properties:
                    data:
                      "$ref": "#/components/schemas/PaymentRecordResponse"
        '400':
          "$ref": "#/components/responses/BadRequest"
        '401':
          "$ref": "#/components/responses/Unauthorized"
        '403':
          "$ref": "#/components/responses/Forbidden"
        '404':
          "$ref": "#/components/responses/NotFound"
        '409':
          "$ref": "#/components/responses/Conflict"
        default:
          "$ref": "#/components/responses/DefaultError"
        '503':
          "$ref": "#/components/responses/ServiceUnavailable"
      operationId: postOrdersByOrderIdPayments
  "/orders/{orderId}/logistics":
    get:
      tags:
      - Bill
      summary: 获取物流信息
      description: 权限：DOCTOR（只读）/ CS / ADMIN
      parameters:
      - name: orderId
        in: path
        required: true
        schema:
          type: integer
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                allOf:
                - "$ref": "#/components/schemas/CommonResponse"
                - type: object
                  properties:
                    data:
                      "$ref": "#/components/schemas/LogisticsResponse"
        '400':
          "$ref": "#/components/responses/BadRequest"
        '401':
          "$ref": "#/components/responses/Unauthorized"
        '403':
          "$ref": "#/components/responses/Forbidden"
        '404':
          "$ref": "#/components/responses/NotFound"
        '409':
          "$ref": "#/components/responses/Conflict"
        default:
          "$ref": "#/components/responses/DefaultError"
        '503':
          "$ref": "#/components/responses/ServiceUnavailable"
      operationId: getOrdersByOrderIdLogistics
    post:
      tags:
      - Bill
      summary: 录入物流信息并发货
      description: |
        权限：CS / ADMIN
        任务 9D.14 第一增量：发货前必须存在该订单最后一道工序节点的 OUT/PASS 终检出检记录；缺失时返回 409。
        录入成功后 external_status 自动变为 SHIPPED，医生端收到 WebSocket 通知。
      parameters:
      - name: orderId
        in: path
        required: true
        schema:
          type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/LogisticsRequest"
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                allOf:
                - "$ref": "#/components/schemas/CommonResponse"
                - type: object
                  properties:
                    data:
                      "$ref": "#/components/schemas/LogisticsResponse"
        '400':
          "$ref": "#/components/responses/BadRequest"
        '401':
          "$ref": "#/components/responses/Unauthorized"
        '403':
          "$ref": "#/components/responses/Forbidden"
        '404':
          "$ref": "#/components/responses/NotFound"
        '409':
          "$ref": "#/components/responses/Conflict"
        default:
          "$ref": "#/components/responses/DefaultError"
        '503':
          "$ref": "#/components/responses/ServiceUnavailable"
      operationId: postOrdersByOrderIdLogistics
  "/logistics/orders":
    get:
      tags:
      - Bill
      summary: 客服配送列表
      description: |
        权限：CS / ADMIN
        任务 9D.91 第一增量：一期人工配送管理列表，可按物流状态筛选；不接真实物流 API、电子面单或物流平台 webhook。
      parameters:
      - name: logistics_status
        in: query
        required: false
        schema:
          type: string
          enum:
          - PENDING
          - SHIPPED
          - EXCEPTION
          - FOLLOWING
          - RESOLVED
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          default: 50
          minimum: 1
          maximum: 200
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                allOf:
                - "$ref": "#/components/schemas/CommonResponse"
                - type: object
                  properties:
                    data:
                      type: array
                      items:
                        "$ref": "#/components/schemas/DeliveryOrderResponse"
        '400':
          "$ref": "#/components/responses/BadRequest"
        '401':
          "$ref": "#/components/responses/Unauthorized"
        '403':
          "$ref": "#/components/responses/Forbidden"
        '404':
          "$ref": "#/components/responses/NotFound"
        '409':
          "$ref": "#/components/responses/Conflict"
        default:
          "$ref": "#/components/responses/DefaultError"
        '503':
          "$ref": "#/components/responses/ServiceUnavailable"
      operationId: getLogisticsOrders
  "/orders/{orderId}/logistics/exception":
    post:
      tags:
      - Bill
      summary: 记录物流异常跟进
      description: |
        权限：CS / ADMIN
        任务 9D.91 第一增量：客服人工维护物流异常状态与内部跟进说明；跟进说明只作为内部客服消息留痕，不返回给医生端物流详情。
      parameters:
      - name: orderId
        in: path
        required: true
        schema:
          type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/LogisticsExceptionRequest"
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                allOf:
                - "$ref": "#/components/schemas/CommonResponse"
                - type: object
                  properties:
                    data:
                      "$ref": "#/components/schemas/DeliveryOrderResponse"
        '400':
          "$ref": "#/components/responses/BadRequest"
        '401':
          "$ref": "#/components/responses/Unauthorized"
        '403':
          "$ref": "#/components/responses/Forbidden"
        '404':
          "$ref": "#/components/responses/NotFound"
        '409':
          "$ref": "#/components/responses/Conflict"
        default:
          "$ref": "#/components/responses/DefaultError"
        '503':
          "$ref": "#/components/responses/ServiceUnavailable"
      operationId: postOrdersByOrderIdLogisticsException
...